// dashboard/src/utils/hasura-api/generated/schemas/createArrayRelationshipBulkOperation.ts
/**
 * Generated by orval v8.1.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateArrayRelationshipBulkOperationType } from './createArrayRelationshipBulkOperationType';
import type { CreateArrayRelationshipStep } from './createArrayRelationshipStep';

export interface CreateArrayRelationshipBulkOperation {
  /** Type of operation to create array relationship */
  type: CreateArrayRelationshipBulkOperationType;
  /** Resource version of the operation */
  resource_version: number;
  args: CreateArrayRelationshipStep[];
}