// dashboard/src/utils/hasura-api/generated/schemas/createObjectRelationshipBulkOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateObjectRelationshipBulkOperationType } from './createObjectRelationshipBulkOperationType';
import type { CreateObjectRelationshipStep } from './createObjectRelationshipStep';

export interface CreateObjectRelationshipBulkOperation {
  /** Type of operation to create object relationship */
  type: CreateObjectRelationshipBulkOperationType;
  /** Resource version of the operation */
  resource_version: number;
  args: CreateObjectRelationshipStep[];
}